Update import paths after moving products folder
