+++ client/package.json
@@ -8,7 +8,6 @@
     "@material-ui/core": "^4.5.0",
     "@material-ui/pickers": "^3.2.6",
     "@material-ui/icons": "^4.4.3",
-    "@material-ui/pickers": "^3.2.6",
     "axios": "^0.19.0",
     "date-fns": "^2.4.1",
     "react": "^16.10.1",
